Add websocket support for the ESP targets.
The used async TCP/Http library also supports Websockets out of the box, no dependencies added.

As per spec, each Thing gets its own Websocket connection.

The property changed message is implemented by having a changed flag in each Thing Property.
The event loop will iterate over all Things and Properties and send a changed-property message if necessary and resets the flag.

Why not directly sending a changed message in the property setter method?
First, because there are pointers back to the ThingAdapter and second the event loop must be the send initiator, because the multicore ESP32 could otherwise cause potential concurrency issues.

An alternative of iterating over all Things and Properties would be a linked list of changed properties.
Because it is expected that only a few Things with only a few Properties are present on such a restricted device, it doesn't matter much though. For one Thing with one Property the runtime cost is the even same.

Signed-off-by: David Gräff <david.graeff@web.de>